feat: Windows webview/package/console signals, platform bin releases
- klain:webview over Edge WebView2 and -package on Windows (ADR-00725/00726)
- ConPTY test harness; tty/tui console branches and signal delivery verified;
  console SIGINT wake fix, SIGBREAK (ADR-00727/00728)
- process.kill by signal name; var re-declaration fix (Test262 S12.14_A11);
  atoi/readlink declare guards (ADR-00729/00730/00731)
- CI matrix on linux x64/arm64, macOS x64/arm64, windows x64; release attaches
  version-stamped binaries per passed platform; --version; install.sh/.ps1
  (TDD-00179, ADR-00732)

+++ codegen/llvm/embedassetssrc/embed_assets.c
@@ -149,8 +149,8 @@ static void kml_embed_handle(int fd, const void *blob) {
   char hdr[512];
   if (kml_embed_resolve(blob, p, rlen, &data, &dlen, &ctype)) {
     int hn = snprintf(hdr, sizeof(hdr),
-      "HTTP/1.0 200 OK\r\nContent-Type: %s\r\nContent-Length: %ld\r\n"
-      "Connection: close\r\n\r\n", kml_ctype_str(ctype), dlen);
+      "HTTP/1.0 200 OK\r\nContent-Type: %s\r\nContent-Length: %lld\r\n"
+      "Connection: close\r\n\r\n", kml_ctype_str(ctype), (long long)dlen);
     send(fd, hdr, (size_t)hn, 0);
     if (!head && dlen > 0) {
       int64_t off = 0;
